modularized the shooting stars logic
